requirements.txt
Flask
praat-parselmouth
